requirements.txt
openai>=1.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
colorama>=0.4.6
